Struct clap_lex::RawOsString

pub struct RawOsString(_);

A container for owned byte strings converted by this crate.

For more information, see RawOsStr.

Implementations

impl RawOsString

pub fn new(string: OsString) -> RawOsString

Converts a platform-native string into a representation that can be more easily manipulated.

For more information, see RawOsStr::new.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
println!("{:?}", RawOsString::new(os_string));

pub fn from_string(string: String) -> RawOsString

Wraps a string, without copying or encoding conversion.

This method is much more efficient than RawOsString::new, since the encoding used by this crate is compatible with UTF-8.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::from_string(string.clone());
assert_eq!(string, raw);

pub fn assert_from_raw_vec(string: Vec<u8, Global>) -> RawOsString

Wraps a byte string, without copying or encoding conversion.

Panics

Panics if the string is not valid for the unspecified encoding used by this crate.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let raw = RawOsString::new(os_string);
let raw_bytes = raw.clone().into_raw_vec();
assert_eq!(raw, RawOsString::assert_from_raw_vec(raw_bytes));

pub unsafe fn from_raw_vec_unchecked(string: Vec<u8, Global>) -> RawOsString

Wraps a byte string, without copying or encoding conversion.

Safety

The string must be valid for the unspecified encoding used by this crate.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let raw = RawOsString::new(os_string);
let raw_bytes = raw.clone().into_raw_vec();
assert_eq!(raw, unsafe {
    RawOsString::from_raw_vec_unchecked(raw_bytes)
});

pub fn clear(&mut self)

Equivalent to String::clear.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let mut raw = RawOsString::new(os_string);
raw.clear();
assert!(raw.is_empty());

pub fn into_box(self) -> Box<RawOsStr, Global>

Equivalent to String::into_boxed_str.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::from_string(string.clone());
assert_eq!(string, *raw.into_box());

pub fn into_os_string(self) -> OsString

Converts this representation back to a platform-native string.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let raw = RawOsString::new(os_string.clone());
assert_eq!(os_string, raw.into_os_string());

pub fn into_raw_vec(self) -> Vec<u8, Global>

Returns the byte string stored by this container.

The returned string will use an unspecified encoding.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::from_string(string.clone());
assert_eq!(string.into_bytes(), raw.into_raw_vec());

pub fn into_string(self) -> Result<String, RawOsString>

Equivalent to OsString::into_string.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::from_string(string.clone());
assert_eq!(Ok(string), raw.into_string());

pub fn shrink_to_fit(&mut self)

Equivalent to String::shrink_to_fit.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let mut raw = RawOsString::from_string(string.clone());
raw.shrink_to_fit();
assert_eq!(string, raw);

pub fn split_off(&mut self, at: usize) -> RawOsString

Equivalent to String::split_off.

Panics

Panics if the index is not a valid boundary.

Examples

use os_str_bytes::RawOsString;

let mut raw = RawOsString::from_string("foobar".to_owned());
assert_eq!("bar", raw.split_off(3));
assert_eq!("foo", raw);

pub fn truncate(&mut self, new_len: usize)

Equivalent to String::truncate.

Panics

Panics if the index is not a valid boundary.

Examples

use os_str_bytes::RawOsString;

let mut raw = RawOsString::from_string("foobar".to_owned());
raw.truncate(3);
assert_eq!("foo", raw);